Move CustomDatePicker to reusable component
