+++ src/views/CampusSafety/views/LostAndFoundAdmin/views/MissingItemList/MissingItemList.module.scss
@@ -18,7 +18,6 @@
   display: flex;
   align-items: center;
   gap: 1rem;
-
   color: var(--mui-palette-primary-contrastText); // Contrast text color for readability
   padding: 1rem;
   border-radius: 5px;
